fixes style
