+++ src/app/Attributes/Structure.php
@@ -10,6 +10,6 @@ class Structure
         'appends', 'auth', 'buttons', 'controls', 'comparisonOperator', 'countCache',
         'crtNo', 'dataRouteSuffix', 'debounce', 'dtRowId', 'flatten', 'fullInfoRecordLimit',
         'lengthMenu', 'method', 'model', 'name',  'preview', 'responsive', 'searchMode',
-        'searchModes', 'selectable', 'templateCache', 
+        'searchModes', 'selectable', 'templateCache',
     ];
 }
