+++ src/app/Services/Data/Filters/BaseFilter.php
@@ -4,8 +4,8 @@ namespace LaravelEnso\Tables\app\Services\Data\Filters;
 
 use Illuminate\Database\Eloquent\Builder;
 use LaravelEnso\Tables\app\Contracts\Table;
-use LaravelEnso\Tables\app\Services\Data\Config;
 use LaravelEnso\Tables\app\Contracts\Filter;
+use LaravelEnso\Tables\app\Services\Data\Config;
 
 
 abstract class BaseFilter implements Filter
