+++ src/app/Services/Template/Validators/Columns.php
@@ -4,7 +4,6 @@ namespace LaravelEnso\Tables\app\Services\Template\Validators;
 
 use LaravelEnso\Helpers\app\Classes\Obj;
 use LaravelEnso\Tables\app\Attributes\Style;
-use LaravelEnso\Tables\app\Services\Template;
 use LaravelEnso\Tables\app\Exceptions\ColumnException;
 use LaravelEnso\Tables\app\Attributes\Column as Attributes;
